GEOMESA-2720 Support AWS credentials controller service
* Add AWSCredentialsProviderService config to HBase and FS processors
* Use nar inheritance correctly
* Separate DataStoreConfigService out into distinct nars
* Move all processors to single module
