+++ geomesa-nifi-controllers/geomesa-nifi-controllers-impl/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -1,0 +1,1 @@
+org.geomesa.nifi.controller.impl.AccumuloDataStoreConfigControllerService
