+++ geomesa-nifi-controllers/pom.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.geomesa.nifi</groupId>
+        <artifactId>geomesa-nifi</artifactId>
+        <version>2.4.0-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>geomesa-nifi-controllers</artifactId>
+    <packaging>pom</packaging>
+
+    <description>Parent module for controller services</description>
+
+    <!-- see https://nifi.apache.org/docs/nifi-docs/html/developer-guide.html#nars
+         for an explanation of the layout of this module -->
+
+    <modules>
+        <module>geomesa-nifi-controllers-api</module>
+        <module>geomesa-nifi-controllers-api-nar</module>
+        <module>geomesa-nifi-controllers-impl</module>
+        <module>geomesa-nifi-controllers-nar</module>
+    </modules>
+
+</project>
