--- geomesa-nifi-nar/pom.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--********************************************************************
-* Copyright (c) 2015-2017 Commonwealth Computer Research, Inc.
-* All rights reserved. This program and the accompanying materials
-* are made available under the terms of the Apache License, Version 2.0 which
-* accompanies this distribution and is available at
-* http://www.opensource.org/licenses/apache2.0.php.
-*********************************************************************-->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.geomesa.nifi</groupId>
-        <artifactId>geomesa-nifi</artifactId>
-        <version>2.4.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>geomesa-nifi-nar</artifactId>
-    <packaging>nar</packaging>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors-accumulo</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors-hbase</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors-kafka</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors-fs</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.geomesa.nifi</groupId>
-            <artifactId>geomesa-nifi-processors-redis</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-tools_2.11</artifactId>
-            <classifier>data</classifier>
-            <version>${geomesa.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.hadoop</groupId>
-            <artifactId>hadoop-aws</artifactId>
-        </dependency>
-
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.nifi</groupId>
-                <artifactId>nifi-nar-maven-plugin</artifactId>
-                <extensions>true</extensions>
-            </plugin>
-        </plugins>
-    </build>
-</project>
