--- geomesa-nifi-processors-fs/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,0 @@
-org.geomesa.nifi.fs.PutGeoMesaFileSystem
