+++ geomesa-nifi-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,2 +1,7 @@
-org.geomesa.nifi.geo.PutGeoTools
-org.geomesa.nifi.geo.ConvertToGeoAvro
+org.geomesa.nifi.processors.datastore.PutGeoTools
+org.geomesa.nifi.processors.ConvertToGeoAvro
+org.geomesa.nifi.processors.datastore.PutGeoMesaAccumulo
+org.geomesa.nifi.processors.datastore.PutGeoMesaFileSystem
+org.geomesa.nifi.processors.datastore.PutGeoMesaHBase
+org.geomesa.nifi.processors.datastore.PutGeoMesaKafka
+org.geomesa.nifi.processors.datastore.PutGeoMesaRedis
